+++ conversor.py
@@ -1,47 +1,13 @@
 import os
 import ffmpeg
-from datetime import datetime
+from src.config import read_config
+from src.log import create_log_file
+from src.utils import get_video_size
 
-def get_video_size(file_path):
-    return os.path.getsize(file_path)
-
-def read_config(config_file):
-    default_config = {
-        'MB_ALVO': 7.925,
-        'ULTIMOS_SEGUNDOS': 30,
-        'MAX_TENTATIVAS': 10,
-        'X': 1280,
-        'Y': 720,
-        'FPS': 25
-    }
-
-    # Verifica se o arquivo de configuração existe
-    if not os.path.exists(config_file):
-        # Cria o arquivo de configuração com os valores padrão
-        with open(config_file, 'w') as f:
-            for key, value in default_config.items():
-                f.write(f"{key} = {value}\n")
-
-    # Lê as configurações do arquivo
-    config = {}
-    with open(config_file, 'r') as f:
-        for line in f:
-            if '=' in line:
-                key, value = line.split('=')
-                config[key.strip()] = float(value.strip())
-
-    return config
-
-def create_log_file(log_file, message):
-    now = datetime.now()
-    timestamp = now.strftime("%Y-%m-%d %H:%M:%S")
-    with open(log_file, 'a') as f:  
-        f.write(f"[{timestamp}] {message}\n")
-
-def compress_and_trim_video(input_file, output_file, target_size_mb, target_last, max_attempts, scale_X, scale_Y, fps, tolerance=0.99):
+def compress_and_trim_video(input_file, output_file, target_size_mb, target_last, max_attempts, scale_X, scale_Y, fps, log_file, tolerance=0.99):
     target_size_bytes = target_size_mb * 1024 * 1024
     min_size_bytes = (target_size_mb * tolerance) * 1024 * 1024
-    create_log_file(log_file,f"target_size_bytes: {target_size_bytes} | min_size_bytes: {min_size_bytes}")
+    create_log_file(log_file, f"target_size_bytes: {target_size_bytes} | min_size_bytes: {min_size_bytes}")
 
     # Obtem informações sobre o vídeo
     probe = ffmpeg.probe(input_file)
@@ -53,7 +19,7 @@ def compress_and_trim_video(input_file, output_file, target_size_mb, target_last
         print(f"ULTIMOS_SEGUNDOS ({target_last}) é maior que a duração do vídeo ({duration:.2f}). Cancelando conversão.")
         return
 
-    # Definane o ponto de início para os últimos segundos do vídeo
+    # Define o ponto de início para os últimos segundos do vídeo
     start_time = duration - target_last
 
     # Define a taxa de bits para atingir o tamanho alvo
@@ -76,7 +42,7 @@ def compress_and_trim_video(input_file, output_file, target_size_mb, target_last
                 'c:a': 'aac',
                 'b:a': '128k',
                 'r': f'{fps}',
-                'vf': 'scale='f'{scale_X}:'f'{scale_Y},unsharp=5:5:1.0:5:5:0.0',
+                'vf': f'scale={scale_X}:{scale_Y},unsharp=5:5:1.0:5:5:0.0',
                 'profile:v': 'high',
                 'level:v': '4.0',
                 'pix_fmt': 'yuv420p'
@@ -87,16 +53,16 @@ def compress_and_trim_video(input_file, output_file, target_size_mb, target_last
         # Verifica o tamanho do arquivo de saída
         output_size_bytes = get_video_size(output_file)
         diff_percent = ((output_size_bytes - target_size_bytes) / target_size_bytes)
-        diff = diff_percent*100
+        diff = diff_percent * 100
         if min_size_bytes <= output_size_bytes <= target_size_bytes or abs(diff) < 0.1:
             create_log_file(log_file, f"diff_percent {diff_percent*100:.2f}%")
             create_log_file(log_file, f"Sucesso! output_size_bytes {output_size_bytes}")
             break
         elif output_size_bytes > target_size_bytes:
-            bit_rate *= 1 - (diff_percent* 1.1)  # Reduz a taxa de bits
+            bit_rate *= 1 - (diff_percent * 1.1)  # Reduz a taxa de bits
             create_log_file(log_file, f"output_size_bytes {output_size_bytes} > target_size_bytes {target_size_bytes} = bit_rate - {diff:.2f}%")
         else:
-            bit_rate *= 1 + (abs(diff_percent)* 1.1)  # Aumenta a taxa de bits
+            bit_rate *= 1 + (abs(diff_percent) * 1.1)  # Aumenta a taxa de bits
             create_log_file(log_file, f"output_size_bytes {output_size_bytes} < min_size_bytes {min_size_bytes} = bit_rate + {diff:.2f}%")
 
         attempts += 1
@@ -105,16 +71,16 @@ def compress_and_trim_video(input_file, output_file, target_size_mb, target_last
         print(f"Limite de tentativas ({max_attempts}) atingido. Não foi possível alcançar o tamanho desejado.")
         create_log_file(log_file, f"Limite de tentativas ({max_attempts}) atingido. Não foi possível alcançar o tamanho desejado.")
 
-def process_videos_in_folder(input_folder, output_folder, config_file):
+def process_videos_in_folder(input_folder, output_folder, config_file, log_file):
     # Lê as configurações do arquivo
     config = read_config(config_file)
     target_size_mb = config.get('MB_ALVO', 7.925)
     target_last = config.get('ULTIMOS_SEGUNDOS', 30)
     max_attempts = config.get('MAX_TENTATIVAS', 10)
-    scale_X = config.get('X', 1920)
-    scale_Y = config.get('Y', 1080)
+    scale_X = config.get('X', 1280)
+    scale_Y = config.get('Y', 720)
     fps = config.get('FPS', 25)
-    create_log_file(log_file,f"Configuração carregada: {config}")
+    create_log_file(log_file, f"Configuração carregada: {config}")
 
     # Verifica se a pasta de entrada existe, se não, cria-a
     if not os.path.exists(input_folder):
@@ -123,7 +89,7 @@ def process_videos_in_folder(input_folder, output_folder, config_file):
         os.makedirs(input_folder)
 
         # Cria o arquivo de instruções
-        with open(('Instrucoes.txt'), 'w') as f:
+        with open(os.path.join(input_folder, 'Instrucoes.txt'), 'w') as f:
             f.write("Instruções de Uso:\n")
             f.write("- Organize seus vídeos na pasta 'Input';\n")
             f.write("- O script converterá e comprimirá os vídeos para a pasta 'Output';\n")
@@ -145,12 +111,13 @@ def process_videos_in_folder(input_folder, output_folder, config_file):
                 output_file = os.path.join(output_folder, f"{os.path.splitext(file)[0]}_Convertido.mp4")
                 
                 # Chama a função para converter e comprimir o vídeo
-                compress_and_trim_video(input_file, output_file, target_size_mb, target_last, max_attempts, scale_X, scale_Y, fps)
+                compress_and_trim_video(input_file, output_file, target_size_mb, target_last, max_attempts, scale_X, scale_Y, fps, log_file)
 
-input_folder = './Input'
-output_folder = './Output'
-config_file = 'Config.txt'
-log_file = os.path.join('log.txt')
+if __name__ == '__main__':
+    input_folder = './Input'
+    output_folder = './Output'
+    config_file = 'Config.txt'
+    log_file = 'log.txt'
 
-create_log_file(log_file, "Iniciando conversor...")
-process_videos_in_folder(input_folder, output_folder, config_file)
+    create_log_file(log_file, "Iniciando conversor...")
+    process_videos_in_folder(input_folder, output_folder, config_file, log_file)
