+++ src/log.py
@@ -1,0 +1,7 @@
+from datetime import datetime
+
+def create_log_file(log_file, message):
+    now = datetime.now()
+    timestamp = now.strftime("%Y-%m-%d %H:%M:%S")
+    with open(log_file, 'a') as f:
+        f.write(f"[{timestamp}] {message}\n")
