+++ src/utils.py
@@ -1,0 +1,4 @@
+import os
+
+def get_video_size(file_path):
+    return os.path.getsize(file_path)
